head params and save model and remove mirror strategy

+++ dataset/general_task.py
@@ -56,11 +56,11 @@ class GeneralTasks:
                 b_bboxes, b_cates, num_bbox = gen_bboxes(
                     self.batch_size, b_coors, b_cates, self.max_obj_num,
                     image_input_sizes)
-                targets['b_bboxes'] = tf.cast(b_bboxes, tf.float32)
+                targets['b_bboxes'] = b_bboxes
                 one_hot = self._one_hots(b_cates, 1)
                 targets['b_cates'] = tf.where(one_hot == 0., np.inf,
                                               one_hot) - 1.
-                targets['num_bbox'] = tf.cast(num_bbox, tf.float32)
+                targets['num_bbox'] = num_bbox
         return tf.cast(new_imgs, dtype=tf.float32), targets
 
     def _resize_coors(self, annos, original_sizes, resize_size,
